requirements.txt
fastapi==0.111.0
sqlalchemy
psycopg2-binary
pyjwt
passlib[brypt]